+++ p2n/util.py
@@ -7,7 +7,9 @@ import logging
 import operator
 import functools
 import itertools
+import traceback
 import subprocess
+from StringIO import StringIO
 from json.encoder import JSONEncoder
 from collections import OrderedDict
 
@@ -120,6 +122,7 @@ class JsonObjectEncoder(JSONEncoder):
     Serialize nested object compositions to JSON
     """
     def default(self, o):
+        # TODO: Maybe use "o.asdict()"?
         return o.__dict__
 
 
@@ -159,3 +162,21 @@ def object_to_dictionary(obj, rules):
             data[key] = value
 
     return data
+
+
+def exception_traceback(exc_info=None):
+    """
+    Return a string containing a traceback message for the given
+    exc_info tuple (as returned by sys.exc_info()).
+
+    from setuptools.tests.doctest
+    """
+
+    if not exc_info:
+        exc_info = sys.exc_info()
+
+    # Get a traceback message.
+    excout = StringIO()
+    exc_type, exc_val, exc_tb = exc_info
+    traceback.print_exception(exc_type, exc_val, exc_tb, file=excout)
+    return excout.getvalue()